feat: Update publisher client to make use of manager client
